+++ lib/features/settings/components/rows/reminder_notification_customize_word.dart
@@ -4,7 +4,7 @@ import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/molecules/premium_badge.dart';
 import 'package:pilll/features/premium_introduction/premium_introduction_sheet.dart';
-import 'package:pilll/features/settings/reminder_notification_customize_word_page.dart';
+import 'package:pilll/features/reminder_notification_customize_word/page.dart';
 import 'package:pilll/entity/setting.codegen.dart';
 
 class ReminderNotificationCustomizeWord extends HookConsumerWidget {
